backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
Pillow==10.2.0
django-filter==23.5
django-storages==1.14.2
django-rest-auth==0.9.5
django-allauth==0.60.1
django-celery-beat==2.5.0
celery==5.3.6
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
requests==2.31.0
stripe==7.11.0
django-redis==5.4.0
django-cleanup==8.0.0
django-guardian==2.4.0
django-ratelimit==4.1.0
django-simple-history==3.4.0
django-import-export==3.3.4
django-debug-toolbar==4.2.0
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1
black==24.1.1
flake8==7.0.0
isort==5.13.2 